Add login and signup forms

+++ controllers/todo.js
@@ -1,9 +1,11 @@
-const { getCurUser } = require('../temp/tmpData');
 const Todo = require('../models/Todo');
 
 exports.getIndex = async(req, res, next) => {
-    const curUser = await getCurUser();
-    return Todo.find({ user: curUser.id }).populate().sort({ createdAt: -1 })
+    const curUser = req.user;
+    if (!curUser) {
+        return res.redirect('/login');
+    }
+    return Todo.find({ user: curUser.id }).populate('user').sort({ createdAt: -1 })
         .then(async todos => {
             userTodos = todos
                 .map(item => ({
@@ -16,21 +18,31 @@ exports.getIndex = async(req, res, next) => {
                 user: curUser
             });
         })
+
 };
 
 exports.createTodo = async(req, res, next) => {
     const { text } = req.body;
-    const curUser = await getCurUser();
+    const curUser = req.user;
+    if (!curUser) {
+        return res.redirect('/login');
+    }
     const todo = new Todo({
         text: text,
         user: curUser
     });
-    todo.save();
-    console.log("Item Saved");
-    return res.redirect(303, '/');
+    const item = await todo.save();
+    if (item) {
+        console.log("Item Saved");
+        return res.redirect(303, '/');
+    }
+    return res.status(500).send("Not Added");
 };
 
 exports.deleteTodo = async(req, res, next) => {
+    if (!req.user) {
+        return res.status(401).json({ message: "Unauthorized!" });
+    }
     const todoId = req.params.id;
     const response = await Todo.findByIdAndDelete(todoId);
     if (!response) {
@@ -41,6 +53,9 @@ exports.deleteTodo = async(req, res, next) => {
 }
 
 exports.updateTodo = async(req, res, next) => {
+    if (!req.user) {
+        return res.status(401).json({ message: "Unauthorized!" });
+    }
     const todoId = req.params.id;
     const text = req.body.text;
     const response = await Todo.findByIdAndUpdate(todoId, { text });
@@ -53,16 +68,18 @@ exports.updateTodo = async(req, res, next) => {
 }
 
 exports.getAllTodos = async(req, res, next) => {
-    return Todo.find().populate().sort({ createdAt: -1 })
+    return Todo.find().populate('user').sort({ createdAt: -1 })
         .then(async todos => {
             userTodos = todos
                 .map(item => ({
                     id: item.id,
-                    text: item.text
+                    text: item.text,
+                    owner: item.user.name
                 }));
             return res.render('allTodos', {
                 title: 'All Todo Items',
-                todos: userTodos
+                todos: userTodos,
+                user: req.user
             });
         })
 }
